Add initial JS for displaying notice and tracking acknowledgment
This is our first attempt at a privacy notice that displays across MIT
Libraries properties. When the user acknowledges the notice, a cookie
is set to store that acknowledgment across the domain, so the user
doesn't see the notice when visiting other properties.

Because we want this to be a standalone solution that's easily injected
into existing and new sites, we decided to use styles from the MIT Libraries
design system as inline CSS. The insertRule() method is used to add styles
that cannot be inserted inline (specifically, those that require pseudo-class
selectors).

Related tickets: ARP-363 and ARP-388